adds messages for no data for some combinations of varriables seleced
for some locations and time periods there is no data, missing data, or -999 for all values. In these cases we need to display an message to the user to let them know that this happend.  The message provides details about how to fix, in future commits we may not even allow the user to select the combinations.  This also creates a blank chart to avoid any confusions.