adds ability to switch between avg as bar and avg as line with yearly
Most charts have summed data for the period as the bars and averages as a trend line. The original and still default does the opposite.  This should allow us to take a look at the more typical way of making the average a line against the yearly  data as the chart