improves hover text, I hope
Hover text was not very descriptive and did not call out correct year ranges. There is still issues. Plotly does a poor job of formating the hover box and providing minimal customization. So it's a bit "hacky" - I had to use space characters to achieve padding—also, the average line, not able to initiate hover when over the entire range.